Change Password and Curl Password web app completed!
At the current stage, the users are able to login into the website and change the password. The 'change password section' is only allowed when the users are signed in. The change password section, also contains an token. In which is necessary in order to change the password and to do the authentification.