All API calls handled, curl and app CRUD for Authentification is DONE.
At the moment, the user are able to login in, change password, see different views and log out of the website. The website is also clearning the inputs for better experience. The next step, will be working on the UI/UX and the board of the tic tact toe. For future commits, the 'game API round 1' will be located in a different branch, named API Round 1.