Game Update Engine works!
At the current stage, I still need to create functionally for the tie option of the game. But the winner is displaying a message for X or O and the board is being displayed on the screen only after the user is sign in. I still have to work with the number of games that the user won, and the reset function.